Fix a refcount leak for the allocated shm object in NaClGioShmAllocCtor.
NaClGioShmCtorIntern takes a reference to the passed-in shm object if
it succeeds, so the NaClGioShmAllocCtor should always unref the shm
object that it created regardless of whether the NaClGioShmCtorIntern
function succeeds or fails.  If it failed, the NaClDescUnref will
remove the last reference and trigger the dtor chain.  If it succeeds,
then ownership of the shm object will have a single reference count
left, i.e., the ownership has been passed into the Gio object.

R=mseaborn@google.com,sehr@google.com
BUG= http://code.google.com/p/nativeclient/issues/detail?id=3179

Review URL: https://codereview.chromium.org/11411252

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10343 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2