x86-32 TLS: Remove the IRT's self-pointer requirement for the thread pointer
I added this check in July in r9152 when I was changing NaCl's %gs
segment to point outside of untrusted address space.

The original intention for the check was to ensure that we could
switch back to the original %gs segment arrangement if we wanted to.
However, the new arrangement has proven itself now, so we wouldn't
need to switch back.

I had originally wanted to extend this check to x86-64 and ARM for
consistency, but ARM's native ABI normally uses the 4 bytes at the
thread pointer for other purposes.  (We made ARM's thread pointer a
self-pointer temporarily between r9925 and r9961, but this was only to
aid the ARM r9 ABI switchover.)

The current check is not portable because it means that a PNaCl pexe
that works on ARM or x86-64 could fail on x86-32.  So, to make things
more uniform, we should remove the check.

Note that removing the check is not required for implementing the
PNaCl TLS ABI cleanup in issue 2837: The ExpandTls LLVM pass still
uses a TLS layout that is compatible with the self-pointer check, so
that it will produce nexes that work on older versions of NaCl.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2837
TEST=none

Review URL: https://codereview.chromium.org/11411310

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10402 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2