Fix atomic_ops_nexe_test not to time out on the Valgrind bot
In this test, if the threads are scheduled in the order they are
launched, the CompareAndSwap() loops will complete quickly.  However,
if they are scheduled out-of-order, these loops could take a long time
to complete.  This seems to be happening after changing newlib's
libpthread to use futexes (r10391), which presumably perturbs the
thread scheduling somewhat.

Fix this by yielding the CPU rather than spinning.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3213
TEST="run_atomic_ops_nexe_test buildbot=memcheck" on trybots

Review URL: https://codereview.chromium.org/11514003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10431 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2