Provide more informative errors if we fail to load the IRT inside Chromium
Don't ignore the error code we get any more.

Increase the logging level of some errors.

Remove the message "The user executable was probably not built to use
the IRT" because it is misleading: the last two instances of this
error that I've seen occurred for other reasons.

BUG=http://code.google.com/p/chromium/issues/detail?id=161776
TEST=none

Review URL: https://codereview.chromium.org/11745019

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10530 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2