Thread suspension: Disallow setting register state during context switch
We already disallow setting register state for a thread that was
suspended while fully inside a syscall.  This change also disallows
setting register state for a thread that was suspended during the
syscall context switch between trusted and untrusted code.

This is necessary on Mac OS X because SetRegisters(natp,
GetRegisters(natp)) was not idempotent there as a result of the
workarounds we use for restoring x86-32 segment registers.

It will also be necessary when we later change
NaClAppThreadGetSuspendedRegisters() to return untrusted register
state if the thread is suspended during a context switch.

This change should fix the GDB test run_gdb_break_continue_thread_test
so that it's no longer flaky on Mac.

We add a stress test that reproduces the problem on Mac outside of the
GDB tests.

The existing NaClSignalContextIsUntrusted() doesn't do what we want on
Mac, so we add a new function and rename the original.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3243
TEST=run_thread_suspension_test

Review URL: https://codereview.chromium.org/11784029

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10552 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2