Move CPU features into its own static library.
CPU features has outgrown its original x86-only roots and is used in
more than just the validators. Move it out of there as a second step to
cleaning up this code.

The eventual goal is for CPU features to have similar interfaces for all
CPUs, as well as for some features to work when host and target are
different (e.g. validate an ARM nexe with a certain ARM feature set on
an x86 box).

This CL doesn't change any code beyond:
- moving it;
- updating the includes/include guards;
- fixing silly build scripts.

R= bsy@chromium.org

Review URL: https://codereview.chromium.org/11864002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10590 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2