[MIPS] Add thread-pointer to data addressing register list.
There is already a rule which prevents modification of thread pointer,
so we can be sure the value inside is correct. Thus, no need to guard
loads and stores via tp.
In addition, the change gets rid of some non-POD static constructors.

The commit queue is not working for NaCl at the moment, so we are committing
with:
NOTRY=true
However, we have run a try job manually, and it passed fine.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=2275
TEST= pnacl/build.sh misc-tools

Review URL: https://chromiumcodereview.appspot.com/11876041

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10686 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2