Add NACL_LIKELY() compiler annotations
This gives a hint to the compiler to optimise code so that the common
case (or at least the case that matters for speed) is straight-line
code in which conditional jumps are not taken.

I am using this sparingly, only on critical paths.  It could increase
code size unnecessarily if added everywhere.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3280
TEST=jump count from run_signal_handler_single_step_test

Review URL: https://codereview.chromium.org/12188005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10712 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2