Mac: support trusted_performance_test.
The performance test in the trusted environment isn't run on the Mac because
the Mac doesn't implement clock_gettime or __thread. The tests only ever call
clock_gettime to obtain the CLOCK_MONOTONIC clock, which is easy to emulate
using other clock routines. With that change, all of the performance tests can
be timed. The TLS test is disabled because it seems pointless to exmaine the
performance of a distinct TLS model.

A trailing comma and the end of an enum list in nacl_clock.h is removed. The
trailing comma is legal under C99 but not C++98/C++03. This change results in
nacl_clock.h being #included into a C++ file for the first time.

TEST=scons run_trusted_performance_test
Review URL: https://codereview.chromium.org/12326139

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@10919 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2