Prevent attempting to load the IRT twice
The command channel has an RPC where the IRT can be specified.  (This
is a "dev" interface.)  Since the service runtime may also want to
load the IRT as a result of the -B flag (in the standalone executable)
or as a result of the NaClChromeMainArgs irt_fd member set to a file
descriptor, this results in the IRT being loaded twice.  This change
records whether the IRT has been loaded by the command channel RPC,
and avoids reloading it again.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3318

Review URL: https://codereview.chromium.org/12945003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11021 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2