PNaCl: Disable the -expand-varargs pass for now
ExpandVarArgs doesn't correctly handle passing a struct as a fixed
argument to a varargs function, e.g.
  void foo(struct big b, ...);

This is because the pass strips off the "byval" argument attribute in
the LLVM IR.

This was discovered via the gcc torture tests.

Disable the pass until it is fixed.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3338
TEST=smoke test: run_hello_world_test

Review URL: https://codereview.chromium.org/12940009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11055 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2