Add compatible mmap to IRT
Our old mmap code ignored PROT_EXEC flag and old versions of
glibc rely on that.  Glibc no longer requires this hack and
we can disable it in latest version of mmap interface but we
need to support it for old binaries.

BUG = https://code.google.com/p/nativeclient/issues/detail?id=3402
TESTS = trybots
R=bsy@google.com

Review URL: https://codereview.chromium.org/14065026

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11243 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2