Conditional linking with libgcc_eh to allow stable bitcode pexes not to expose symbols needed by it.
[see issue https://code.google.com/p/nativeclient/issues/detail?id=3069 for a
full discussion]

Some symbols (malloc, free, abort, strlen) are exposed by the pexe because
libgcc_eh references them. Since we don't support exceptions in the first
public stable ABI for PNaCl, there's no need to link the libgcc_eh library.

This CL makes the linking of libgcc_eh conditional. By default it is not
being linked, and the above symbols are not exposed by the pexe. This is done
by removing them from the crt1.x linker script that sets them to exported.

To make this work, some other dependencies on libgcc_eh have to be resolved.
crtbegin is compiled conditionally to have a version that doesn't depend on
libgcc_eh for stable ABI builds. In addition, a stub implementation of
_Unwind_* functions is provided for this case.

The driver scripts (pnacl-clang[++] and pnacl-translate) are adjusted to
support the stable (no EH) build by default, and provide a EH-supporting build
when the --pnacl-allow-exceptions flag is provided. pnacl-clang already had
that flag, but for pnacl-translate it's new.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3069

Review URL: https://codereview.chromium.org/14553003

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11271 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2